Fix word addressing
Now uses for loop instead of goto

Also fixed bug -- malloc was not allocating enough memory because we
were using the same MEM_SIZE for malloc and IP loop even though IP was
word-width and needed to be incremented differently in the loop. changed
malloc to use MEM_SIZE * sizeof(Word) now so that IP goes from 0 to
65535 (i.e., max representable by WORD_SIZE) and accesses values with a
granularity of word size.

Also added ASan to Makefile